The three things a timer needs:
is a timer object that notifies a  object or class after a certain
amount of specified time passes.
The second thing a timer need is an ActionListener that does a certain
action every time the specified time passes in the time object.
Lastly a start method to initiate the timer to start.